Modularize Block Factory
