+++ modules/block/factory/go.mod
@@ -1,0 +1,5 @@
+module github.com/treeverse/lakefs/modules/block/factory
+
+go 1.23
+
+// This module uses the go.work file to get all package dependencies from lakefs
