+++ modules/block/factory/go.work
@@ -1,0 +1,5 @@
+go 1.23
+
+use (
+	.
+)
