+++ pkg/block/factory/build.go
@@ -27,15 +27,6 @@ const (
 )
 
 func BuildBlockAdapter(ctx context.Context, statsCollector stats.Collector, c params.AdapterConfig) (block.Adapter, error) {
-	adapter, err := buildBlockAdapter(ctx, statsCollector, c)
-	if err != nil {
-		return nil, err
-	}
-
-	return block.NewMetricsAdapter(adapter), nil
-}
-
-func buildBlockAdapter(ctx context.Context, statsCollector stats.Collector, c params.AdapterConfig) (block.Adapter, error) {
 	blockstore := strings.ToLower(c.BlockstoreType())
 	logging.FromContext(ctx).
 		WithField("type", blockstore).
